[Numerics] remove Warning

+++ numerics/src/tools/line_search.c
@@ -76,7 +76,7 @@ void get_non_monotone_ref(void* nm_ref_data, double* theta_ref)
     case NM_LS_MAX: // classical nonmonotone theta_ref = max theta_j
       data_max = (nm_ref_struct*) nm_ref_data;
       local_theta_ref = *theta_ref;
-      for (size_t i = 0; i < data_max->m; ++i)
+      for (int i = 0; i < data_max->m; ++i)
       {
         if (data_max->previous_thetas[i] > local_theta_ref)
         {
@@ -90,7 +90,7 @@ void get_non_monotone_ref(void* nm_ref_data, double* theta_ref)
       data_mean = (nm_ref_struct*)nm_ref_data;
       if (data_mean->m > 0)
       {
-        for (size_t i = 0; i < data_mean->m; ++i)
+        for (int i = 0; i < data_mean->m; ++i)
         {
           local_theta_ref += data_mean->previous_thetas[i];
         }
